complete authentication

+++ demo/src/main/java/com/javaweb/repository/MediaItemRepository.java
@@ -3,5 +3,5 @@ package com.javaweb.repository;
 import com.javaweb.entity.MediaItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MediaItemRepository extends JpaRepository<MediaItemEntity,Long> {
+public interface MediaItemRepository extends JpaRepository<MediaItemEntity,Integer> {
 }
